refactor: Reorganize files into webapp structure
- Moved ChatGPT component to components/webapp/
- Moved WebRTC hook to hooks/webapp/
- Moved agent instructions to prompts/webapp/
- Moved API endpoints to internal routes
- Created redirects for backward compatibility
- Updated imports and tests
- All tests passing
